Moved all other booking services to separate namespaces
