+++ Api/Services/Accommodations/Bookings/Documents/IBookingDocumentsService.cs
@@ -5,7 +5,7 @@ using HappyTravel.Edo.Api.Models.Bookings;
 using HappyTravel.Edo.Api.Models.Payments;
 using HappyTravel.Edo.Data.Documents;
 
-namespace HappyTravel.Edo.Api.Services.Accommodations.Bookings
+namespace HappyTravel.Edo.Api.Services.Accommodations.Bookings.Documents
 {
     public interface IBookingDocumentsService
     {
