+++ Api/Services/Accommodations/Bookings/Management/BookingPermissionHelper.cs
@@ -2,7 +2,7 @@ using HappyTravel.Edo.Api.Models.Agents;
 using HappyTravel.Edo.Common.Enums;
 using HappyTravel.Edo.Data.Bookings;
 
-namespace HappyTravel.Edo.Api.Services.Accommodations.Bookings
+namespace HappyTravel.Edo.Api.Services.Accommodations.Bookings.Management
 {
     public static class BookingPermissionHelper
     {
