+++ Api/Services/Accommodations/Bookings/Payments/IAppliedBookingMarkupRecordsManager.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using HappyTravel.Edo.Api.Models.Markups;
 
-namespace HappyTravel.Edo.Api.Services.Accommodations.Bookings
+namespace HappyTravel.Edo.Api.Services.Accommodations.Bookings.Payments
 {
     public interface IAppliedBookingMarkupRecordsManager
     {
