+++ Api/Services/Accommodations/Bookings/ResponseProcessing/BookingResponseProcessor.cs
@@ -1,11 +1,12 @@
 using System.Threading.Tasks;
 using HappyTravel.Edo.Api.Infrastructure.Logging;
 using HappyTravel.Edo.Api.Models.Users;
+using HappyTravel.Edo.Api.Services.Accommodations.Bookings.Management;
 using HappyTravel.EdoContracts.Accommodations;
 using HappyTravel.EdoContracts.Accommodations.Enums;
 using Microsoft.Extensions.Logging;
 
-namespace HappyTravel.Edo.Api.Services.Accommodations.Bookings
+namespace HappyTravel.Edo.Api.Services.Accommodations.Bookings.ResponseProcessing
 {
     public class BookingResponseProcessor : IBookingResponseProcessor
     {
