+++ Api/Services/Accommodations/Bookings/ResponseProcessing/IBookingResponseDataLogService.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using HappyTravel.EdoContracts.Accommodations;
 
-namespace HappyTravel.Edo.Api.Services.Accommodations.Bookings
+namespace HappyTravel.Edo.Api.Services.Accommodations.Bookings.ResponseProcessing
 {
     public interface IBookingAuditLogService
     {
